+++ NEWS.md
@@ -1,5 +1,6 @@
 # smoothr 0.0.1.9000
 
+- Remove holes in polygons with `fill_holes()`
 - Drop small lines and polygons with `drop_crumbs()`
 - Kernel smoothing implemented in `smooth_ksmooth()`
 - Densification of spatial lines and polygons implemented with `densify()`